refactor function into methods
