+++ nilearn/interfaces/bids/glm.py
@@ -2,7 +2,6 @@
 
 import inspect
 import json
-import warnings
 from pathlib import Path
 
 import numpy as np
@@ -10,54 +9,6 @@ import numpy as np
 from nilearn import __version__
 from nilearn._utils import logger
 from nilearn._utils.glm import coerce_to_dict
-from nilearn.externals import tempita
-
-
-def _clean_contrast_name(contrast_name):
-    """Remove prohibited characters from name and convert to camelCase.
-
-    .. versionadded:: 0.9.2
-
-    BIDS filenames, in which the contrast name will appear as a
-    contrast-<name> key/value pair, must be alphanumeric strings.
-
-    Parameters
-    ----------
-    contrast_name : :obj:`str`
-        Contrast name to clean.
-
-    Returns
-    -------
-    new_name : :obj:`str`
-        Contrast name converted to alphanumeric-only camelCase.
-    """
-    new_name = contrast_name[:]
-
-    # Some characters translate to words
-    new_name = new_name.replace("-", " Minus ")
-    new_name = new_name.replace("+", " Plus ")
-    new_name = new_name.replace(">", " Gt ")
-    new_name = new_name.replace("<", " Lt ")
-
-    # Others translate to spaces
-    new_name = new_name.replace("_", " ")
-
-    # Convert to camelCase
-    new_name = new_name.split(" ")
-    new_name[0] = new_name[0].lower()
-    new_name[1:] = [c.title() for c in new_name[1:]]
-    new_name = " ".join(new_name)
-
-    # Remove non-alphanumeric characters
-    new_name = "".join(ch for ch in new_name if ch.isalnum())
-
-    # Let users know if the name was changed
-    if new_name != contrast_name:
-        warnings.warn(
-            f'Contrast name "{contrast_name}" changed to "{new_name}"',
-            stacklevel=4,
-        )
-    return new_name
 
 
 def _generate_model_metadata(out_file, model):
@@ -232,10 +183,7 @@ def save_glm_to_bids(
     if prefix and not prefix.endswith("_"):
         prefix += "_"
 
-    verbose = model.verbose
-
     contrasts = coerce_to_dict(contrasts)
-
     for k, v in contrasts.items():
         if not isinstance(k, str):
             raise ValueError(f"contrast names must be strings, not {type(k)}")
@@ -246,31 +194,23 @@ def save_glm_to_bids(
                 f"not {type(v)}"
             )
 
-    if model.__str__() == "Second Level Model":
-        sub_directory = "group"
-    else:
-        sub_directory = (
-            prefix.split("_")[0] if prefix.startswith("sub-") else ""
-        )
-
     out_dir = Path(out_dir)
     out_dir.mkdir(exist_ok=True, parents=True)
 
     dset_desc_file = out_dir / "dataset_description.json"
     _generate_dataset_description(dset_desc_file, model.__str__())
 
-    out_dir = out_dir / sub_directory
-    out_dir.mkdir(exist_ok=True, parents=True)
+    output = model._generate_filenames_output(
+        prefix, contrasts, contrast_types, out_dir
+    )
+    output["dir"].mkdir(exist_ok=True, parents=True)
 
-    # Write out design matrices to files.
+    verbose = model.verbose
+
     if hasattr(model, "design_matrices_"):
         design_matrices = model.design_matrices_
     else:
         design_matrices = [model.design_matrix_]
-
-    output = _generate_filenames_output(
-        prefix, design_matrices, contrasts, contrast_types, out_dir
-    )
 
     logger.log("Generating design matrices figures...", verbose=verbose)
     # TODO: Assuming that cases of multiple design matrices correspond to
@@ -287,13 +227,13 @@ def save_glm_to_bids(
 
         # Save design matrix and associated figure
         design_matrix.to_csv(
-            out_dir / f"{prefix}{run_str}design.tsv",
+            output["dir"] / f"{prefix}{run_str}design.tsv",
             sep="\t",
             index=False,
         )
 
         if model.__str__() == "First Level Model":
-            with (out_dir / f"{prefix}{run_str}design.json").open(
+            with (output["dir"] / f"{prefix}{run_str}design.json").open(
                 "w"
             ) as f_obj:
                 json.dump(
@@ -305,7 +245,7 @@ def save_glm_to_bids(
 
     # Model metadata
     # TODO: Determine optimal mapping of model metadata to BIDS fields.
-    metadata_file = out_dir / f"{prefix}statmap.json"
+    metadata_file = output["dir"] / f"{prefix}statmap.json"
     _generate_model_metadata(metadata_file, model)
 
     logger.log(
@@ -316,77 +256,16 @@ def save_glm_to_bids(
         for output_type in contrast_maps:
             img = statistical_maps[contrast_name][output_type]
             filename = output["statistical_maps"][contrast_name][output_type]
-            img.to_filename(out_dir / filename)
+            img.to_filename(output["dir"] / filename)
 
     logger.log("Saving contrast-level statistical maps...", verbose=verbose)
-    _write_model_level_statistical_maps(model, prefix, out_dir)
+    _write_model_level_statistical_maps(model, prefix, output["dir"])
 
     logger.log("Generating HTML...", verbose=verbose)
     glm_report = model.generate_report(
         contrasts=contrasts, input=output, verbose=verbose - 1, **kwargs
     )
-    glm_report.save_as_html(out_dir / f"{prefix}report.html")
-
-
-def _generate_filenames_output(
-    prefix, design_matrices, contrasts, contrast_types, out_dir
-):
-    design_matrices_dict = tempita.bunch()
-    contrasts_dict = tempita.bunch()
-    for i_run, _ in enumerate(design_matrices, start=1):
-        run_str = f"run-{i_run}_" if len(design_matrices) > 1 else ""
-
-        design_matrices_dict[i_run] = tempita.bunch(
-            design_matrix=f"{prefix}{run_str}design.svg",
-            correlation_matrix=f"{prefix}{run_str}corrdesign.svg",
-        )
-
-        tmp = {
-            contrast_name: (
-                f"{prefix}{run_str}"
-                f"contrast-{_clean_contrast_name(contrast_name)}"
-                "_design.svg"
-            )
-            for contrast_name in contrasts
-        }
-        contrasts_dict[i_run] = tempita.bunch(**tmp)
-
-    if not isinstance(contrast_types, dict):
-        contrast_types = {}
-
-    statistical_maps = {}
-    for contrast_name in contrasts:
-        # Extract stat_type
-        contrast_matrix = contrasts[contrast_name]
-
-        # Strings and 1D arrays are assumed to be t-contrasts
-        if isinstance(contrast_matrix, str) or (contrast_matrix.ndim == 1):
-            stat_type = "t"
-        else:
-            stat_type = "F"
-
-        # Override automatic detection with explicit type if provided
-        stat_type = contrast_types.get(contrast_name, stat_type)
-
-        # Convert the contrast name to camelCase
-        contrast_entity = f"contrast-{_clean_contrast_name(contrast_name)}_"
-        suffix = "_statmap.nii.gz"
-        statistical_maps[contrast_name] = {
-            "effect_size": (f"{prefix}{contrast_entity}stat-effect{suffix}"),
-            "stat": (f"{prefix}{contrast_entity}stat-{stat_type}{suffix}"),
-            "effect_variance": (
-                f"{prefix}{contrast_entity}stat-variance{suffix}"
-            ),
-            "z_score": (f"{prefix}{contrast_entity}stat-z{suffix}"),
-            "p_value": (f"{prefix}{contrast_entity}stat-p{suffix}"),
-        }
-
-    return {
-        "dir": out_dir,
-        "design_matrices_dict": design_matrices_dict,
-        "contrasts_dict": contrasts_dict,
-        "statistical_maps": statistical_maps,
-    }
+    glm_report.save_as_html(output["dir"] / f"{prefix}report.html")
 
 
 def _write_model_level_statistical_maps(model, prefix, out_dir):
